services/core/pyproject.toml
[tool.poetry]
name = "core-service"
version = "0.1.0"
description = "Core Service for Tasks and Purchases"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
aio-pika = "^9.4.0"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
